requirements.txt
cassandra-driver>=3.29,<4
colorama